+++ Joss_Paper/paper.md
@@ -10,9 +10,9 @@ authors:
     orcid: 0000-0003-2791-0648
 #    affiliation: 1
 affiliations:
-  - name: North Carolina State University, Raleigh NC, USA
-#    index: 1
-date: 18 July 2023
+  - name: C. T. Kelley, North Carolina State University, Raleigh NC, USA
+    index: 1
+date: 15 April 2024
 bibliography: paper.bib
 ---
 
