### 2024-04-15 #6621 @ctkelley
On Mon, Apr 15, 2024 at 8:48 AM Daniel S. Katz ***@***.***>
wrote:

> 👋 @ctkelley <https://github.com/ctkelley> - note that your paper does
> not compile. Please follow the example paper
> <https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography>
> and note that you can click on the error above to find out more about it.
> In addition, you could work on the possibly missing (and/or invalid) DOIs
> that editorialbot suggests, but note that some may be incorrect. Please
> feel free to make changes to your .bib file, then use the command @editorialbot
> check references to check again, and the command @editorialbot generate
> pdf after making changes to the .md file or when the references are right
> to make a new PDF. editorialbot commands need to be the first entry in a
> new comment.
>
