### 2024-04-15 #6621 @ctkelley
On Mon, Apr 15, 2024 at 8:58 AM The Open Journals editorial robot <
***@***.***> wrote:

> ⚠️ An error happened when generating the pdf
> <https://github.com/openjournals/joss-papers/actions/runs/8691376307?check_suite_focus=true>.
> Author (C. T. Kelley) is missing affiliation.
>
