+++ Joss_Paper/paper.md
@@ -8,10 +8,10 @@ tags:
 authors:
   - name: C. T. Kelley
     orcid: 0000-0003-2791-0648
-#    affiliation: 1
+    affiliation: 1
 affiliations:
-  - name: C. T. Kelley, North Carolina State University, Raleigh NC, USA
-#    index: 1
+  - name: North Carolina State University, Raleigh NC, USA
+    index: 1
 date: 15 April 2024
 bibliography: paper.bib
 ---
