### 2024-04-15 #6621 @ctkelley
> Try this
>
> ```
> authors:
>   - name: C. T. Kelley
>     orcid: 0000-0003-2791-0648
>     affiliation: 1
> affiliations:
>   - name: North Carolina State University, Raleigh NC, USA
>     index: 1
> ```

Thanks, That did it. Now I will get to work on the DOIs.
