### 2024-04-15 #6621 @danielskatz
@ctkelley - thanks for your submission to JOSS, and your work getting it to build. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy. For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!
