+++ Joss_Paper/paper.bib
@@ -29,7 +29,7 @@ volume = {40},
 number = {2},
 pages = {A817-A847},
 year = {2018},
-doi = {10.1137/17M1140819},
+doi = {10.1137/17M1140819}
 }
 
 @article{dongarra:1983,
@@ -58,6 +58,7 @@ url="http://www.alanturing.net/turing_archive/archive/l/l10/l10.php"
   address = "Philadelphia, PA, USA",
   year = 1996,
   pages = "xxviii+688",
+  doi = {10.1137/1.9780898718027},
   URL = "http://www.ma.man.ac.uk/~higham/asna.html",
   isbn = "0-89871-355-2"
         }
@@ -82,6 +83,7 @@ publisher="SIAM",
 address="Philadelphia",
 isbn="978-1-611977-26-4",
 series="Fundamentals of Algorithms",
+doi={10.1137/1.9781611977271},
 number=20
 }
 
@@ -140,7 +142,8 @@ title="Error Bounds from Extra-Precise Iterative Refinement",
 author="James Demmel and Yozo Hida and William Kahan",
 year=2006,
 journal="ACM Trans. Math. Soft.",
-pages="325--351"
+pages="325--351",
+doi={10.1145/1141885.1141894},
 } 
 
 @article{VanderVorst:1992,
@@ -160,6 +163,7 @@ volume=13
    journal = {SIAM J. Sci. Stat. Comp.},
    year = 1986,
    pages = {856--869},
+   doi = {10.1137/0907058},
    volume = 7   }
 
 }  
