### 2024-04-15 #6621 @ctkelley
On Mon, Apr 15, 2024 at 9:24 AM The Open Journals editorial robot <
***@***.***> wrote:

> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1137/23M1549079 is OK
> - 10.1137/17M1140819 is OK
> - 10.1137/20M1342902 is OK
> - 10.5281/zenodo.4284807 is OK
> - 10.5281/zenodo.7521427 is OK
> - 10.1137/141000671 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Improving the accuracy of computed eigenvalues and...
> - No DOI given, and none found for title: Progress Report on the Automatic Computing Engine
> - 10.1137/1.9780898718027 may be a valid DOI for title: Accuracy and Stability of Numerical Algorithms
> - 10.1137/1.9781611977271 may be a valid DOI for title: Solving Nonlinear Equations with Iterative Methods...
> - 10.61208/pjo-2023-030 may be a valid DOI for title: Newton’s Method in Three Precisions
> - No DOI given, and none found for title: Using MultiPrecisonArrays.jl: Iterative Refinement...
> - 10.1145/1141885.1141894 may be a valid DOI for title: Error Bounds from Extra-Precise Iterative Refineme...
> - No DOI given, and none found for title: Bi-CGSTAB: A fast and smoothly converging variant ...
> - 10.1137/0907058 may be a valid DOI for title: GMRES a generalized minimal residual algorithm for...
>
> INVALID DOIs
>
> - 10.1137/17M112291 is INVALID
>
>
