+++ Joss_Paper/paper.bib
@@ -18,7 +18,7 @@ volume=39,
 number=6,
 pages={A2834-A2856}, 
 year=2017,
-doi={10.1137/17M112291} 
+doi={10.1137/17M112291}
 }
 
 @article{CarsonHigham:2018,
@@ -105,6 +105,7 @@ howpublished="https://github.com/ctkelley/SIAMFANLEquations.jl"
       eprint={2307.16051},
       archivePrefix={arXiv},
       primaryClass={math.NA},
+      doi = {10.48550/arXiv.2307.16051},
       note="To appear in Pacific Journal of Optimization"
 }
 
@@ -124,6 +125,7 @@ title="Using {MultiPrecisonArrays.jl}: {I}terative Refinement in {J}ulia",
 year=2024,
 archivePrefix={arXiv},
 primaryClass={math.NA},
+doi = {10.48550/arXiv.2311.14616},
 eprint={2311.14616}
 }
 
