+++ Joss_Paper/paper.bib
@@ -18,7 +18,7 @@ volume=39,
 number=6,
 pages={A2834-A2856}, 
 year=2017,
-doi={10.1137/17M112291}
+doi={10.1137/17M1122918}
 }
 
 @article{CarsonHigham:2018,
