### 2024-04-15 #6621 @ctkelley
On Mon, Apr 15, 2024 at 12:55 PM The Open Journals editorial robot <
***@***.***> wrote:

> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1137/23M1549079 is OK
> - 10.1137/17M1122918 is OK
> - 10.1137/17M1140819 is OK
> - 10.1137/1.9780898718027 is OK
> - 10.1137/20M1342902 is OK
> - 10.1137/1.9781611977271 is OK
> - 10.5281/zenodo.4284807 is OK
> - 10.48550/arXiv.2307.16051 is OK
> - 10.5281/zenodo.7521427 is OK
> - 10.48550/arXiv.2311.14616 is OK
> - 10.1137/141000671 is OK
> - 10.1145/1141885.1141894 is OK
> - 10.1137/0907058 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Improving the accuracy of computed eigenvalues and...
> - No DOI given, and none found for title: Progress Report on the Automatic Computing Engine
> - No DOI given, and none found for title: Bi-CGSTAB: A fast and smoothly converging variant ...
>
> INVALID DOIs
>
> - None
>
>
