+++ docs/src/MPArray.bib
@@ -204,3 +204,11 @@ year=2006,
 journal="ACM Trans. Math. Soft.",
 pages="325--351"
 }
+
+@misc{accelerate,
+key="Julia Package",
+title="AppleAccelerate.jl",
+note="Julia Package",
+year=2023,
+howpublished="https://github.com/JuliaLinearAlgebra/AppleAccelerate.jl"
+}
