+++ Joss_Paper/paper.bib
@@ -121,7 +121,7 @@ howpublished="https://github.com/ctkelley/MultiPrecisionArrays.jl"
 
 @misc{kelley:2024b,
 author="C. T. Kelley",
-title="Using {MultiPrecisonArrays.jl}: {I}terative Refinement in {J}ulia",
+title="Using {MultiPrecisionArrays.jl}: {I}terative Refinement in {J}ulia",
 year=2024,
 archivePrefix={arXiv},
 primaryClass={math.NA},
