+++ src/Solvers/IRTriangle.jl
@@ -3,7 +3,7 @@ IRTriangle!(AF::Union{MPLFact,MPLEFact,MPHFact}, r, rs, verbose)
 This is the solve phase using the factorization object (AFS) you get
 from a multiprecision LU factorization.
 
-Solve for the defect by quering on_the_fly
+Solve for the defect by querying on_the_fly
 to figure out if we can do the triangular solves entirely in low precision.
 
 If on_the_fly(AF) == false, then demote the residual and solve in low
