+++ dev/MPArray.bib
@@ -141,7 +141,7 @@ pages="A2815--A2835"
 
 @misc{ctk:mparraysdocs,
 author="C. T. Kelley",
-title="Using {MultiPrecisonArrays.jl}: {I}terative Refinement in {J}ulia",
+title="Using {MultiPrecisionArrays.jl}: {I}terative Refinement in {J}ulia",
 year=2024,
 archivePrefix={arXiv},
 primaryClass={math.NA},
