+++ CITATION.bib
@@ -1,6 +1,6 @@
 @misc{ctk:mparraysdocs,
 author="C. T. Kelley",
-title="Using {MultiPrecisonArrays.jl}: {I}terative Refinement in {J}ulia",
+title="Using {MultiPrecisionArrays.jl}: {I}terative Refinement in {J}ulia",
 year=2024,
 archivePrefix={arXiv},
 primaryClass={math.NA},
