+++ docs/src/MPArray.bib
@@ -66,8 +66,8 @@ number=20
 
 @article{skeel,
 author="R. D. Skeel",
-title="Iterative refinement implies numerical stability for 
-{G}aussian elimination",
+title="{Iterative refinement implies numerical stability for 
+Gaussian elimination}",
 journal="Math. Comp.", 
 volume= 35, 
 year=1980, 
