+++ dev/MPArray.bib
@@ -199,7 +199,8 @@ volume=13
 
 @article{demmelir,
 title="Error Bounds from Extra-Precise Iterative Refinement",
-author="James Demmel and Yozo Hida and William Kahan",
+author="James Demmel and Yozo Hida and William Kahan and X. S. Li and
+S. Mukherjee and E. J. Riedy",
 year=2006,
 journal="ACM Trans. Math. Soft.",
 pages="325--351"
