+++ docs/src/MPArray.bib
@@ -188,7 +188,7 @@ year=1983
 }
 
 @article{bicgstab,
-author="H. A. {van der Vorst}",
+author="H. A. van der Vorst",
 title="{Bi-CGSTAB}: A fast and smoothly converging variant to
 {Bi-CG} for the solution of nonsymmetric systems",
 journal = {SIAM J. Sci. Stat. Comp.},
