+++ Joss_Paper/paper.bib
@@ -149,7 +149,7 @@ doi={10.1145/1141885.1141894},
 } 
 
 @article{VanderVorst:1992,
-author="H. A. {van der Vorst}",
+author="H. A. van der Vorst",
 title="{Bi-CGSTAB}: A fast and smoothly converging variant to
 {Bi-CG} for the solution of nonsymmetric systems", 
 journal = {SIAM J. Sci. Stat. Comp.},
