### 2024-04-26 #6621 @vissarion
Hi @vlc1 @imciner2 @Jgoldfar @henrik227 @ChrisRackauckas @rafaelbailo would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
