### 2024-04-30 #6698 @editorialbot
[REVIEW]: MultiPrecisionArrays.jl: A Julia package for iterative refinement
**Submitting author:** @ctkelley (<a href="http://orcid.org/0000-0003-2791-0648">Carl Kelley</a>)
**Repository:** https://github.com/ctkelley/MultiPrecisionArrays.jl
**Branch with paper.md** (empty if default branch): Joss_Paper
**Version:** 0.1.3
