### 2024-06-07 e8b18ed
Update termination criterion (small resid -> small normwise backward error)
