### 2024-06-07 a65ce5d
Update docs for termination criterion change
