+++ docs/src/MPArray.bib
@@ -9,6 +9,16 @@ year = {2024},
 doi = {10.1137/23M1549079}
 }
 
+@article{higham97,
+  author = "N. J. Higham",
+  title = "Iterative refinement for linear systems and {LAPACK}",
+  year = 1997,
+  journal = "IMA J. Numer. Anal.",
+  volume = 17,
+  pages = "495--509"
+}
+
+
 @article{CarsonHigham1,
 author = "E. Carson and N. J. Higham",
 title = {A new analysis of iterative refinement and its application of
