+++ test/Wilkinson/wilk_test.jl
@@ -22,7 +22,7 @@ mout = hires_sol(A, b)
 #
 lres=length(mout.rhist)
 ndel=norm(mout.sol - xe,Inf)
-l1ok=(lres == 4)
+l1ok=(lres <= 4)
 s1ok=(ndel < 1.e-10);
 A1ok=l1ok && s1ok
 A1ok || println("32-32 Failure for alpha=$alpha:   ", lres,"  ",ndel)
@@ -32,7 +32,7 @@ mout = hires_sol(A, b; TF=Float16)
 lres=length(mout.rhist)
 ndel=norm(mout.sol - xe,Inf)
 (alpha < 500.0) ? lt=6 : lt=8
-l2ok=(lres == lt) 
+l2ok=(lres <= lt) 
 s2ok=(ndel < 1.e-10);
 A2ok=l2ok && s1ok
 A2ok || println("32-16 Failure for alpha=$alpha:   ", lres,"  ",ndel)
