+++ docs/src/MPArray.bib
@@ -13,7 +13,7 @@ doi = {10.1137/23M1549079}
   author = "N. J. Higham",
   title = "Iterative refinement for linear systems and {LAPACK}",
   year = 1997,
-  journal = "IMA J. Numer. Anal.",
+  journal = "IMA J. Numerical Anal.",
   volume = 17,
   pages = "495--509"
 }
