### 2024-06-13 c1e733d
Test new termination criterion
