+++ src/Solvers/mpkrir.jl
@@ -116,7 +116,7 @@ function mpkrir(AF::MPKFact, b; reporting = false,
     normtype = Inf
     TB = eltype(b)
     # remember that eps(TB) = 2 * unit roundoff
-    tolf = TB(10.0)*eps(TB)
+    tolf = TB(0.5)*eps(TB)
     n = length(b)
     onetb = TB(1.0)
     bsc = copy(b)
