+++ .github/workflows/ci.yml
@@ -26,9 +26,9 @@ jobs:
           - '~1.11.0-0'
           - 'nightly'
         os:
-#          - ubuntu-latest
+          - ubuntu-latest
           - macOS-latest
-#          - windows-latest
+          - windows-latest
         arch:
           - x64
     steps:
