+++ test/Greens/gtest.jl
@@ -29,7 +29,8 @@ function greensHok(n = 31)
     ok32 = (e32 < 1.e-13) && (l32 <= 4)
     ok32 || println("GreensH fail at TF=Float64-32, $ok32, $e32, $l32")
     (e16, l16) = gtestH(G; TF = Float16)
-    ok16 = (e16 < 1.e-14) && (l16 == 6)
+    ok16 = (e16 < 1.e-14) && (l16 <= 6)
+    println(e16,"  ",l16)
     ok16 || println("GreensH fail at TF=Float64-16")
     (e3216, l3216) = gtestH(G; TF = Float16, TW = Float32)
     ok3216 = (e3216 < 1.e-6) && (l3216 <= 4)
