+++ README.md
@@ -37,7 +37,9 @@ __The half precision LU for Float16 in this package is much faster (more than 10
    - Explanation for why I am not excited about evaluating the residual in extended precision + a bit of support for that anyhow
    - Replacing Polyester with [OhMyThreads](https://github.com/JuliaFolds2/OhMyThreads.jl) v0.5 or later. I am worried about [this](https://discourse.julialang.org/t/why-is-loopvectorization-deprecated/109547/74).
   
- - v0.1.1: Better docs and ...
+ - v0.1.1: Better docs and updated termination criterion (normwise backward error)
+   
+ - v0.1.2: Even better docs and ...
    - Nonlinear solver applications 
    - Krylov-IR for high precision residuals
    
