+++ README.md
@@ -6,9 +6,9 @@
 <!--- [![MultiPrecisionArrays Downloads](https://shields.io/endpoint?url=https://pkgs.genieframework.com/api/v1/badge/MultiPrecisionArrays)](https://pkgs.genieframework.com?packages=MultiPrecisionArrays) --->
 
 
-# MultiPrecisionArrays.jl v0.1.1
+# MultiPrecisionArrays.jl v0.1.2
 
-##  v0.1.1 will have new features but no breaking changes from v0.1.0, the current release.
+##  v0.1.2 will have new features but no breaking changes from v0.1.1, the current release.
 
 ## [C. T. Kelley](https://ctk.math.ncsu.edu)
 
