+++ dev/MPArray.bib
@@ -223,3 +223,14 @@ note="Julia Package",
 year=2023,
 howpublished="https://github.com/JuliaLinearAlgebra/AppleAccelerate.jl"
 }
+
+@misc{ctk:irnote,
+author="C. T. Kelley",
+title="Interprecision transfers in iterative refinement",
+year = 2024, 
+archivePrefix={arXiv},
+primaryClass={math.NA},
+eprint={submit/5701252},
+note="submitted for publication"
+}
+
