### 2024-07-03 5ac0445
Update termination for Krylov-IR
