### 2024-07-03 510c939
Fixing allocation issue
