### 2024-07-04 6524f1a
more allocation fixes
