### 2024-07-04 fdf7ab5
More allocation fixes
