### 2024-07-04 5dc5f7b
Update docs and internals of  factorizations
