### 2024-07-06 76cd600
Getting residual precision in Krylov-IR codes. Docs + CI coming soon.
