### 2024-07-07 94f521e
Removing dead code
