### 2024-07-08 ff61711
Make CI faster
