### 2024-07-08 4a8d5ab
Documenting IR-BiCGSTAB in the docstrings
