+++ v0.1.2/MPArray.bib
@@ -1,0 +1,236 @@
+@article{amestoy:2024,
+author = {Amestoy, Patrick and Buttari, Alfredo and Higham, Nicholas J. and L’Excellent, Jean-Yves and Mary, Theo and Vieubl\'{e}, Bastien},
+title = {Five-Precision GMRES-Based Iterative Refinement},
+journal = {SIAM Journal on Matrix Analysis and Applications},
+volume = {45},
+number = {1},
+pages = {529-552},
+year = {2024},
+doi = {10.1137/23M1549079}
+}
+
+@article{higham97,
+  author = "N. J. Higham",
+  title = "Iterative refinement for linear systems and {LAPACK}",
+  year = 1997,
+  journal = "IMA J. Numerical Anal.",
+  volume = 17,
+  pages = "495--509"
+}
+
+
+@article{CarsonHigham1,
+author = "E. Carson and N. J. Higham",
+title = {A new analysis of iterative refinement and its application of
+accurate solution of ill-conditioned sparse linear systems},
+journal = {SIAM Journal on Scientific Computing},
+volume=39,
+number=6,
+pages={A2834-A2856},
+year=2017,
+doi={10.1137/17M112291}
+}
+
+@article{CarsonHigham,
+author = "E. Carson and N. J. Higham",
+title = {Accelerating the Solution of Linear Systems by Iterative Refinement in Three Precisions},
+journal = {SIAM Journal on Scientific Computing},
+volume = {40},
+number = {2},
+pages = {A817-A847},
+year = {2018},
+doi = {10.1137/17M1140819},
+}
+
+@book{higham,
+  author = "N. J. Higham",
+  title = "Accuracy and Stability of Numerical Algorithms",
+  publisher = "Society for Industrial and Applied Mathematics",
+  address = "Philadelphia, PA, USA",
+  year = 1996,
+  pages = "xxviii+688",
+  URL = "http://www.ma.man.ac.uk/~higham/asna.html",
+  isbn = "0-89871-355-2"
+        }
+
+@article{ctk:sirev20,
+author="C. T. Kelley",
+title="Newton's Method in Mixed Precision",
+year=2022,
+volume=64,
+pages="191--211",
+doi="10.1137/20M1342902",
+journal="SIAM Review"
+}
+
+@book{ctk:fajulia,
+author="C. T. Kelley",
+title="{Solving Nonlinear Equations with Iterative Methods:
+Solvers and Examples in Julia}",
+year=2022,
+publisher="SIAM",
+address="Philadelphia",
+series="Fundamentals of Algorithms",
+number=20
+}
+
+@article{skeel,
+author="R. D. Skeel",
+title="{Iterative refinement implies numerical stability for 
+Gaussian elimination}",
+journal="Math. Comp.", 
+volume= 35, 
+year=1980, 
+pages="817-–832",
+doi="10.1090/S0025-5718-1980-0572859-4"
+}
+
+@article{highamscaling,
+author="N. J. Higham and S. Pranesh and M. Zounon",
+title="Squeezing a matrix into half precision, with
+an application to solving linear systems", 
+journal="SIAM J. Sci. Comp.",
+year=2019, 
+volume=41,
+pages="A2536--A2551"
+}
+
+@misc{ctk:Newton3p,
+      title={Newton's Method in Three Precisions}, 
+      author="C. T. Kelley",
+      year={2023},
+      eprint={2307.16051},
+      archivePrefix={arXiv},
+      primaryClass={math.NA}
+}
+
+@misc{newblas,
+title="A Proposal for a Next-Generation {BLAS}",
[…]
