+++ Joss_Paper/paper.bib
@@ -169,3 +169,14 @@ volume=13
    volume = 7   }
 
 }  
+
+@misc{ctk:irnote,
+author="C. T. Kelley",
+title="Interprecision transfers in iterative refinement",
+year = 2024,
+archivePrefix={arXiv},
+primaryClass={math.NA},
+eprint={2407.00827},
+note="submitted for publication" 
+}
+
