### 2024-07-10 #6698 @ctkelley
Hi, I would like to add a new reference to the paper and a line  or two
about that new work. If the paper is under review, I can wait to do that
until the review is complete. Otherwise I will update the paper.

Please let me know.

Thanks,

— Tim

On Sat, Jun 15, 2024 at 9:06 AM Daniel S. Katz ***@***.***>
wrote:

> 👋 @@vlc1 <https://github.com/vlc1> & @imciner2
> <https://github.com/imciner2> - I just wanted to check on if you are
> still able to work on this? (As track editor, I try to check on reviews in
> the CSISM track where no progress has been recorded in a 2-week period.)
>
