### 2024-07-10 #6698 @danielskatz
@vissarion should answer this as the editor, but in my opinion, it shouldn't be a big deal to do
