### 2024-07-11 #6698 @ctkelley
Thanks. I will hold off until I hear from @vissarion

On Wed, Jul 10, 2024 at 11:58 AM Daniel S. Katz ***@***.***>
wrote:

> @vissarion <https://github.com/vissarion> should answer this as the
> editor, but in my opinion, it shouldn't be a big deal to do
>
