+++ Joss_Paper/paper.bib
@@ -170,7 +170,7 @@ volume=13
 
 }  
 
-@misc{ctk:irnote,
+@misc{kelley:2024c,
 author="C. T. Kelley",
 title="Interprecision transfers in iterative refinement",
 year = 2024,
