### 2024-07-16 #6698 @vlc1
Dear @ctkelley, thank for this nice addition to Julia's linear algebra ecosystem. I was able to reproduce the example you provide in the JOSS manuscript:
```julia
using BenchmarkTools
using MultiPrecisionArrays
using MultiPrecisionArrays.Examples

N = 4096
G = Gmat(N)
A = I - G
x = ones(N)
b = A * x

@btime lu($A)
@btime mplu($A)

AF = lu(A); @btime $AF \ $b
MPF = mplu(A); @btime $MPF \ $b

```
which resulted in
```
# lu(A)
  241.440 ms (4 allocations: 128.03 MiB)
# mplu(A)
  120.972 ms (15 allocations: 64.09 MiB)
# AF \ b
  7.387 ms (2 allocations: 32.05 KiB)
# MPF \ b
  236.096 ms (35 allocations: 16.80 KiB)

```

The factorisation is indeed twice faster, but the solution stage is much slower, the number of allocations is significant (support for the in-place `ldiv!` would be worthwhile, imho). Is this expected?

Also, while the README.md and JOSS manuscript clearly stipulate that MultiPrecisionArrays.jl targets dense matrices, I couldn't help try it on a sparse matrix,
```julia
using LinearAlgebra
using SparseArrays
using BenchmarkTools
using MultiPrecisionArrays

using Base: OneTo

laplacian(n) = spdiagm(-1 => -ones(n-1), 0 => 2ones(n), +1 => -ones(n-1))
laplacian(n, m) = kron(I(m), laplacian(n)) + kron(laplacian(m), I(n))

n, m = 32, 32

A = laplacian(n, m)
x = reshape([i * j / (n * m) for i in OneTo(n), j in OneTo(m)], :)
b = A * x

@btime lu($A)
@btime mplu($A)

AF = lu(A); @btime \($AF, $b)
MPF = mplu(A); @btime \($MPF, $b)

```
which led to
```
# lu(A)
  1.492 ms (78 allocations: 1.38 MiB)
# mplu(A)
  7.088 s (25 allocations: 1.40 MiB)
# AF \ b
  22.826 μs (2 allocations: 16.25 KiB)
# MPF \ b
  268.626 μs (30 allocations: 12.83 KiB)

```

While I wasn't expecting this to perform well out of the box, the performance gap in the factorization stage surprised me. Could you also comment on this? Thank you very much!
