### 2024-07-16 #6698 @ctkelley
Thanks @vlc1 for the insightful comments.

I will respond to the two questions separately to avoid one lone
confusing reply.

-- Tim

1) Sparse solvers:

The sparse solvers in Julia only support Float64 and iterative
refinement is unlikely to happen until they do. So when you try
to factor a Float32 sparse matrix, it gets promoted to Float64
before the factorization.

```
julia> using LinearAlgebra

julia> using SparseArrays

julia> using BenchmarkTools

julia> laplacian(n) = spdiagm(-1 => -ones(n-1), 0 => 2ones(n), +1 => -ones(n-1));

julia> n=100; L=laplacian(n);

julia> LF=lu(L);

julia> LF
SparseArrays.UMFPACK.UmfpackLU{Float64, Int64}
L factor:
100×100 SparseMatrixCSC{Float64, Int64} with 199 stored entries:

julia> LS=Float32.(L);

julia> LSF=lu(LS);

julia> eltype(LSF)
Float64

julia> @btime lu($L);
  27.083 μs (83 allocations: 119.40 KiB)

julia> @btime lu($LS);
  27.333 μs (91 allocations: 125.18 KiB)

```
So you should not expect any improvement from doing the LU in single.
mplu builds the MPArray structure and factors the low precision copy.
Since lu will promote that low precision copy, I'd expect the cost of
all that to be high. I did not expect it to be as high as it was, but have not tuned
the code for sparsity at all.

There is an open issue about this here

https://github.com/ctkelley/MultiPrecisionArrays.jl/issues?q=is%3Aopen+is%3Aissue

Moreover, sparse solvers spend a lot of time in graph computations which
are integer computations. That work does not depend on the storage
precision. There has been some work IR for sparse problems, but it's
been in C++ and Julia does not support what that work did.
