### 2024-07-16 #6698 @ctkelley
Hi @vlc1. I really appreciate your finding the problem with the solver time.

In this particular case, the solver takes 5 iteraitons with a call to ldiv! for the low precision factorization and a call to mul! for the matrix-vector product to compute the residual. Each of those is N^2 work, so I'd expect the solve to be (roughly) 10-15 times the cost of ldiv! for the high precision factorization.

I have been able to duplicate your results and there is clearly a performance bug in the solvers. I will fix that and update the repo and let you know when everything is done.

-- Tim
