### 2024-07-16 #6698 @ctkelley
Hi @vic1

I found the problem. When I updated the termination criterion I had to
compute
```
optnorm(A, Inf)
```
which is a very bad idea and costs as much as a few IR iterations. Changing
Inf to 1 makes things a lot better and reverting to termination on small
relative residuals makes it better still.

So, I have to get this sorted out and add a one or two kwargs to mplu, ...
to let the solvers know what do to. I will let you know once this is in
there and documented.

— Tim

On Tue, Jul 16, 2024 at 3:14 PM Vincent Le Chenadec <
***@***.***> wrote:

> Dear @ctkelley <https://github.com/ctkelley>, thanks for the detailed
> answer, that sounds great!
>
