### 2024-07-17 8b2c07a
Performance bug fix: part 1
