+++ README.md
@@ -9,7 +9,8 @@
 # MultiPrecisionArrays.jl v0.1.3
 
 ##  v0.1.3 will have better performance but no breaking changes from v0.1.2, the current release. Performance fix in the works. Here is a list ...
-- The termination criteria will change and I will add some options to make it more/less expensive. Computing __opnorm(A,Inf)__ for the termination criteria costs as much as a couple IR iterations
+- The termination criteria will change and I will add some options to make it more/less expensive. Computing __opnorm(A,Inf)__ for the termination criteria costs as much as a couple IR iterations. So that will become an option and terminating on small residuals will be the default. I will also use $\| A \|_1$ because
+it is the least expensive norm to compute. Further economy will be in v0.1.4.
 - The AppleAccelerate BLAS is slower in triangular sovles than OPEN BLAS by a factor of over two (Float32) and about 1.5 (Float64). I am abandoning Apple's BLAS for IR until this is fixed.
   
 ## [C. T. Kelley](https://ctk.math.ncsu.edu)
@@ -195,7 +196,7 @@ julia> ze=norm(z-x,Inf); zr=norm(b-A*z,Inf)/norm(b,Inf);
 julia> we=norm(w-x,Inf); wr=norm(b-A*w,Inf)/norm(b,Inf);
 
 julia> println("Errors: $ze, $we. Residuals: $zr, $wr")
-Errors: 1.33227e-15, 7.41629e-14. Residuals: 1.33243e-15, 7.40609e-14
+Errors: 2.22045e-16, 6.71685e-14. Residuals: 2.22045e-16, 6.71685e-14
 ```
 
 So the results are equally good.
@@ -209,10 +210,10 @@ is 50% more than ```lu```.
 
 ```
 julia> @belapsed mplu($A)
-8.60945e-02
+8.59528e-02
 
 julia> @belapsed lu!(AC) setup=(AC=copy($A))
-1.42840e-01
+1.42112e-01
 
 ```
 It is no surprise that the factorization in single precision took roughly half as long as the one in double. In the double-single precision case, iterative refinement is a great
@@ -222,10 +223,10 @@ The advantages of IR increase as the dimension increases. IR is less impressive 
 julia> N=30; A=I + Gmat(N); 
 
 julia> @belapsed mplu($A)
-4.19643e-06
+5.22217e-06
 
 julia> @belapsed lu!(AC) setup=(AC=copy($A))
-3.70825e-06
+3.64062e-06
 ```
 
 ### A few details
@@ -273,22 +274,23 @@ julia> # Use \ with reporting=true
 julia> mpout=\(MPF, b; reporting=true);
 
 julia> norm(b-A*mpout.sol, Inf)
-1.33227e-15
+2.22045e-16
 
 julia> # Now look at the residual history
 
 julia> mpout.rhist
-5-element Vector{Float64}:
- 9.99878e-01
- 1.21892e-04
- 5.25805e-11
- 2.56462e-14
- 1.33227e-15
+6-element Vector{Float64}:
+ 1.00000e+00
+ 1.89553e-05
+ 4.56056e-11
+ 3.08642e-14
+ 4.44089e-16
+ 2.22045e-16
 ```
 As you can see, IR does well for this problem. The package uses an initial
 iterate of $x = 0$ and so the initial residual is simply $r = b$
 and the first entry in the residual history is $|| b ||_\infty$. The
-iteration terminates successfully after four matrix-vector products.
+iteration terminates successfully after five matrix-vector products.
 
 You may wonder why the residual after the first iteration was so much
 larger than single precision roundoff. The reason is that the default 
[…]
