### 2024-07-18 585acd9
Performance bug fix: part 2
