### 2024-07-18 10d6cac
Performance bug fix: part 3
