+++ src/Factorizations/mpglu!.jl
@@ -1,5 +1,5 @@
 """
-mpglu!(MPGA::MPGArray)
+mpglu!(MPGA::MPGArray; residterm=residtermdefault)
 Factor a MPGArray using the allocations from the structure.
 
 This function factors the low precision copy
@@ -9,6 +9,11 @@ storage for ```basissize``` Kylov vectors and some other things
 GMRES needs.
 You get a factorization
 object as output and can use ```\\``` to solve linear systems.
+
+The kwarg ```residterm``` sets the termination criterion. 
+```residterm == true``` (default) terminates the iteration on 
+small residuals.  ```residterm == false``` terminates the iteration on
+small normwise backward errors. Look at the docs for details.
 """
 function mpglu!(MPGA::MPGArray; residterm=residtermdefault)
 AL=MPGA.AL
