+++ src/Solvers/mpkrir.jl
@@ -68,6 +68,11 @@ which does the BiCGSTAB-IR solve.
 
 You can also use the ```\\``` operator to harvest iteration statistics.
 
+You may not get exactly the same results for this example on
+different hardware, BLAS, versions of Julia or this package. 
+I am still playing with the termination criteria and the iteration
+count could grow or shrink as I do that.
+
 ## Example
 ```jldoctest
 julia> using MultiPrecisionArrays.Examples
