+++ src/Factorizations/mplu!.jl
@@ -130,6 +130,10 @@ The kwarg ```residterm``` sets the termination criterion.
 small residuals.  ```residterm == false``` terminates the iteration on
 small normwise backward errors. Look at the docs for details.
 
+You may not get exactly the same results for this example on
+different hardware, BLAS, versions of Julia or this package. 
+I am still playing with the termination criteria and the iteration
+count could grow or shrink as I do that.
 
 ## Example
 ```jldoctest
