+++ docs/src/Details/Termination.md
@@ -1,8 +1,13 @@
 # Terminating the while loop
+Todays values are
+
+```math
+C_e = .5, C_r = 1.0
+```
 
 We terminate the loop when 
 ```math
-\| r \| < \tau (\| b \| + \| A \| \| x \|)
+\| r \| < C_e \tau (\| b \| + \| A \| \| x \|)
 ```
 where we use $\tau = 0.5 * eps(TW)$, where $eps(TW)$ is the working
 precision floating
@@ -15,4 +20,7 @@ terminate the iteration if
 \| r_{new} \| \ge .9 \| r_{old} \|
 ```
 even if the small residual condition is not satisfied.
+
+I am still playing with the termination criteria and the iteration
+counts and timings could grow or shrink as I do that.
 
