+++ README.md
@@ -46,7 +46,8 @@ __The half precision LU for Float16 in this package is much faster (more than 10
    - Krylov-IR for high precision residuals
 
 - v0.1.3: Still better docs and ..
-   - Fixing a performance bug. 
+   - Fixing a performance bug.
+   - Add options to termination criterion. Change default to small residuals.
 
 ##  Can I complain about this package?
 
