### 2024-07-20 0ca1f83
Performance bug fix: part 4
