+++ src/Solvers/termination_settings.jl
@@ -2,13 +2,28 @@ function termination_settings(TR, residterm)
 #
 # I am continually tweaking this stuff
 #
-Cr=1.0; Ce=.5;
+# I will terminate IR using either small relative residual
+# (residterm = true, the default)
+# || r || < Cr eps(TR) || b || 
+#
+# or small normwise backward error
+# (residterm = false)
+# || r || < Ce eps(TR) (|| A || || x || + || b ||) 
+#
+# I use the L1 norm for A because that is less expensive that 
+# Linfty.  || A || is as expensive as a few IR iterations, 
+# so it's not the default.
+#
+# I also stop when I see stagnation, ie
+# || r_new || > redmax || r_old ||
+#
+Cr=10.0; Ce=.5; 
+Rr=.1; Re=.1;
+residterm ? redmax = Rr : redmax = Re
 residterm ? tf=Cr : tf=Ce
 tolf = eps(TR)*tf
 #
-#   I'm using the L1 norm because it's much faster.
-#
-residterm ?  anrm = 0.0 : anrm = opnorm(AD, 1)
-term_out = (tolf=tolf, anrm=anrm)
+anrm=0.0
+term_out = (tolf=tolf, anrm=anrm, redmax=redmax)
 return term_out
 end
