+++ docs/src/Details/Termination.md
@@ -19,6 +19,13 @@ or a small relative normwise backward error
 \| r \| < C_e u_r (\| b \| + \| A \| \| x \|)
 ```
 
+Termination on small residual is the default because computing $\| A \|$
+is $N^2$ work is more expensive that a few IR iterations. I am using
+$\| A \|_1$ because that takes less time (data locality) than 
+$\| A \|_\infty$ but it is still a pain. The reason $C_r > C_e$ is
+that putting $\| A \|$ in the termination criterion is very useful
+and making $C_r$ large helps compensate for that.
+
 The problem with these criteria is
 that IR can stagnate, especially for ill-conditioned problems, before
 the termination criterion is attained. We detect stagnation by looking
