+++ docs/make.jl
@@ -38,8 +38,8 @@ makedocs(
         "Home"=>"index.md",
         "Half Precision and Krylov-IR" => Any["Half_1.md",], 
         "More than you want to know" => 
-           Any["Details/Termination.md", "Details/Interprecision_1.md",
-                "Details/Extended.md",],
+           Any["Details/Termination.md", "Details/N2Work.md",
+         "Details/Interprecision_1.md", "Details/Extended.md",],
         "MPArray Constructors" => 
            Any["functions/MPArray.md","functions/MPGArray.md",
                "functions/MPBArray.md",],
