+++ Joss_Paper/paper.bib
@@ -1,182 +1,0 @@
-@article{amestoy:2024,
-author = {Amestoy, Patrick and Buttari, Alfredo and Higham, Nicholas J. and L’Excellent, Jean-Yves and Mary, Theo and Vieubl\'{e}, Bastien},
-title = {Five-Precision GMRES-Based Iterative Refinement},
-journal = {SIAM Journal on Matrix Analysis and Applications},
-volume = {45},
-number = {1},
-pages = {529-552},
-year = {2024},
-doi = {10.1137/23M1549079}
-}
-
-@article{CarsonHigham:2017,
-author = {E. Carson and N. J. Higham},
-title = {A new analysis of iterative refinement and its application of
-accurate solution of ill-conditioned sparse linear systems},
-journal = {SIAM Journal on Scientific Computing},
-volume=39,
-number=6,
-pages={A2834-A2856}, 
-year=2017,
-doi={10.1137/17M1122918}
-}
-
-@article{CarsonHigham:2018,
-author = {E. Carson and N. J. Higham},
-title = {Accelerating the Solution of Linear Systems by Iterative Refinement in Three Precisions},
-journal = {SIAM Journal on Scientific Computing},
-volume = {40},
-number = {2},
-pages = {A817-A847},
-year = {2018},
-doi = {10.1137/17M1140819}
-}
-
-@article{dongarra:1983,
-author="J.J.Dongarra and C.B.Moler and J.H.Wilkinson",
-title="Improving the accuracy of computed eigenvalues and eigenvectors",
-journal="SIAM Journal on Numerical Analysis",
-volume=20,
-pages="23--45",
-year=1983
-}
-
-@techreport{Wilkinson:48,
-author="J. H. Wilkinson",
-title="Progress Report on the Automatic Computing Engine",
-number="MA/17/1024",
-institution="Mathematics Division, Department of Scientific and Industrial Research, National Physical Laboratory, Teddington, UK",
-year=1948,
-url="http://www.alanturing.net/turing_archive/archive/l/l10/l10.php"
-}
-
-
-@book{higham:1996,
-  author = "N. J. Higham",
-  title = "Accuracy and Stability of Numerical Algorithms",
-  publisher = "Society for Industrial and Applied Mathematics",
-  address = "Philadelphia, PA, USA",
-  year = 1996,
-  pages = "xxviii+688",
-  doi = {10.1137/1.9780898718027},
-  URL = "http://www.ma.man.ac.uk/~higham/asna.html",
-  isbn = "0-89871-355-2"
-        }
-
-
-@article{kelley:2022a,
-author="C. T. Kelley",
-title="Newton's Method in Mixed Precision",
-year=2022,
-volume=64,
-pages="191--211",
-doi="10.1137/20M1342902", 
-journal="SIAM Review"
-}
-
-@book{kelley:2022b,
-author="C. T. Kelley",
-title="{Solving Nonlinear Equations with Iterative Methods:
-Solvers and Examples in Julia}",
-year=2022,
-publisher="SIAM",
-address="Philadelphia",
-isbn="978-1-611977-26-4",
-series="Fundamentals of Algorithms",
-doi={10.1137/1.9781611977271},
-number=20
-}
-
-@misc{kelley:2022c,
-title="{SIAMFANLEquations.jl}",
-author="C. T. Kelley",
-year=2022,
-note="Julia Package",
-doi="10.5281/zenodo.4284807",
-url="https://github.com/ctkelley/SIAMFANLEquations.jl",
-howpublished="https://github.com/ctkelley/SIAMFANLEquations.jl"
-}
-
-
-@misc{kelley:2023,
[…]
