+++ Joss_Paper/paper.md
@@ -1,289 +1,0 @@
----
-title: 'MultiPrecisionArrays.jl: A Julia package for iterative refinement'
-tags:
-  - Julia
-  - mathematics
-  - numerical analysis
-  - linear equations
-authors:
-  - name: C. T. Kelley
-    orcid: 0000-0003-2791-0648
-    affiliation: 1
-affiliations:
-  - name: North Carolina State University, Raleigh NC, USA
-    index: 1
-date: 15 April 2024
-bibliography: paper.bib
----
-
-# Summary
-
-[MultiPrecisionArrays.jl](https://github.com/ctkelley/MultiPrecisionArrays.jl),
-[@kelley:2024a; @kelley:2024b] provides data structures and
-solvers for several variations of iterative refinement (IR). IR can speed up an LU matrix factorization for solving linear systems of equations by
-factoring a low precision copy of the matrix and using that 
-low precision factorization in a iteration loop to solve the system.
-For example, if high precision is double and low precision is single,
-then the factorization time is cut in half.
-The additional storage cost is the low precision copy, 
-so IR is at time vs storage trade off. IR has a long history
-and a good account of the classical theory is in [@higham:1996].
-
-# Statement of need
-
-The solution of linear systems of equations is a ubiquitous task
-in computational science and engineering. A common method for dense
-systems is Gaussian elimination done via an LU factorization,
-[@higham:1996].  Iterative refinement is a way to reduce the factorization
-time at the cost of additional storage. 
-[MultiPrecisionArrays.jl](https://github.com/ctkelley/MultiPrecisionArrays.jl)
-enables IR with a simple interface in Julia [@Juliasirev:17] with an
-IR factorization object that one uses in the same way as the one for LU.
-The package offers several variants of IR, both classical
-[@Wilkinson:48; @higham:1996], and some from the recent literature
-[@CarsonHigham:2017; @amestoy:2024].
-
-# Algorithm
-
-This package will make solving dense systems of linear equations faster by using the LU factorization and IR. While other factorizations can be used in IR, the
-package is limited to LU for now. A very generic description of this 
-for solving a linear system $A x = b$ in a high (working) precision is
-
-__IR(A, b)__
-
-- $x = 0$
-
-- $r = b$
-
-- Factor $A = LU$ in a lower precision
-
-- While $\| r \|$ is too large
-
-  - $d = (LU)^{-1} r$
-
-  - $x = x + d$
-
-  - $r = b - Ax$
-
-- end
-
-- end
-
-
-In Julia, a code to do this would solve the linear system $A x = b$ in 
-in the working precision, say double,
-by using a factorization in a lower (factorization) 
-precision, say single, within a residual correction iteration. 
-This means that one would need to allocate storage for a copy of $A$ 
-in the factorization precision and factor that copy. 
-
-The multiprecision factorization ```mplu``` makes the low precision copy
-of the matrix, factors that copy, and allocates some storage for the 
-iteration. The original matrix and the low precision factorization
-are stored in a factorization object that you can use with ```\```.
-
-IR is a perfect example of a storage/time trade off.
-To solve a linear system $A x = b$ in $R^N$ with IR,
-one incurs the storage penalty of making a low
-precision copy of $A$ and reaps the benefit of only having to
-factor the low precision copy.
-
-# Installation
-
-The standard way to install a package is to type ```import.Pkg; Pkg.add("MultiPrecisionArrays")``` at the Julia prompt. One can run the unit tests with ```Pkg.test("MultiPrecisionArrays")```.
[…]
