### 2024-07-24 a0344b0
Update paper with solve data
