### 2024-07-25 #6698 @ctkelley
Hi, @vlc1

I have fixed the performance bug. Thanks again for finding that.

I have also updated the paper a bit with some timing data for the complete solve on lines 121 - 140 of the draft.

I made more substantial changes to the documentation to explain the issues
with [termination criteria](https://ctkelley.github.io/MultiPrecisionArrays.jl/dev/Details/Termination)
and the [relative costs of the factorization and the triangular solves](https://ctkelley.github.io/MultiPrecisionArrays.jl/dev/Details/N2Work).

If the new version of the paper is acceptable, I will register v0.1.3 with the performance bug fix and the updated documentation included.

-- Tim
