+++ Project.toml
@@ -14,7 +14,7 @@ LinearAlgebra = "1"
 SIAMFANLEquations = "1"
 SparseArrays = "1"
 julia = "1"
-OhMyThreads = "0.5.0"
+OhMyThreads = "0.5"
 Test = "1"
 
 [extras]
