### 2024-08-14 #6698 @vissarion
@vlc1 are you OK with the fix described above by @ctkelley ?

@imciner2 are you still able to work on this review?

@danielskatz  I just pinged the reviewers and waiting for responses. Since this is a holiday season in several places in Europe (Greece included, where I am located) I expect a bit of delay in responses.
