### 2024-09-20 #6698 @vissarion
@rafaelbailo thanks a lot, please follow the instructions in the beginning of this thread
https://github.com/openjournals/joss-reviews/issues/6698#issue-2271161297
