### 2024-09-23 #6698 @rafaelbailo
Hi @ctkelley (CC: @vissarion), I have just submitted my review comments as an [issue](https://github.com/ctkelley/MultiPrecisionArrays.jl/issues/51) in the package's repository. Please let me know when you get the chance to address them, or if you'd like to discuss any of them further.
