+++ src/Structs4MP/MPGArray.jl
@@ -41,16 +41,15 @@ the Krylov basis, and a few other things GMRES needs. If the high precision
 matrix is double, the low precision is single by default. Half is an option
 which you get with TF=Float16.
 """
-function MPGArray(AH::AbstractArray{Float64,2}; TR=nothing,
-                 basissize=10, TF=Float32)
-AL=TF.(AH)
-(m,n)=size(AH)
-(TR == nothing) ? TRR=eltype(AH) : TRR = TR
-res=ones(TRR,n)
-sol=ones(TRR,n)
-VStore=zeros(TRR,n,basissize)
-KStore=KIRstore(n,"gmres")
-MPGA=MPGArray(AH, AL, VStore, KStore, res, sol, true)
+function MPGArray(AH::AbstractArray{Float64,2}; TR = nothing, basissize = 10, TF = Float32)
+    AL = TF.(AH)
+    (m, n) = size(AH)
+    (TR == nothing) ? TRR = eltype(AH) : TRR = TR
+    res = ones(TRR, n)
+    sol = ones(TRR, n)
+    VStore = zeros(TRR, n, basissize)
+    KStore = KIRstore(n, "gmres")
+    MPGA = MPGArray(AH, AL, VStore, KStore, res, sol, true)
 end
 
 
@@ -62,17 +61,16 @@ the Krylov basis, and a few other things GMRES needs. Since High precision is
 single, low is half. I'm leaving the kwarg for TF in there because it makes
 is easier to cut/paste calls to MPGArray different precisions into a CI loop.
 """
-function MPGArray(AH::AbstractArray{Float32,2}; basissize=10, TF=Float16,
-                  TR=nothing)
-AL=TF.(AH)
-(m,n)=size(AH)
-(TR == nothing) ? TRR=eltype(AH) : TRR = TR
-res=ones(TRR,n)
-sol=ones(TRR,n)
-VStore=zeros(TRR,n,basissize)
-KStore=KIRstore(n,"gmres")
-MPGA=MPGArray(AH, AL, VStore, KStore, res, sol, true)
-return MPGA
+function MPGArray(AH::AbstractArray{Float32,2}; basissize = 10, TF = Float16, TR = nothing)
+    AL = TF.(AH)
+    (m, n) = size(AH)
+    (TR == nothing) ? TRR = eltype(AH) : TRR = TR
+    res = ones(TRR, n)
+    sol = ones(TRR, n)
+    VStore = zeros(TRR, n, basissize)
+    KStore = KIRstore(n, "gmres")
+    MPGA = MPGArray(AH, AL, VStore, KStore, res, sol, true)
+    return MPGA
 end
 
 """     
@@ -80,18 +78,17 @@ KIRstore(n, lsolver, TR=Float64)
 
 Preallocates the vectors a Krylov method uses internally.
 """
-function KIRstore(n, lsolver, TR=Float64)
-    tmp1 = zeros(TR,n)
-    tmp2 = zeros(TR,n)
-    tmp3 = zeros(TR,n)
-    tmp4 = zeros(TR,n)
+function KIRstore(n, lsolver, TR = Float64)
+    tmp1 = zeros(TR, n)
+    tmp2 = zeros(TR, n)
+    tmp3 = zeros(TR, n)
+    tmp4 = zeros(TR, n)
     if lsolver == "gmres"
         return (tmp1, tmp2, tmp3, tmp4)
     else
-        tmp5 = zeros(TR,n)
-        tmp6 = zeros(TR,n)
-        tmp7 = zeros(TR,n)
+        tmp5 = zeros(TR, n)
+        tmp6 = zeros(TR, n)
+        tmp7 = zeros(TR, n)
         return (tmp1, tmp2, tmp3, tmp4, tmp5, tmp6, tmp7)
     end
 end
-
