+++ src/Solvers/IRTriangle.jl
@@ -13,17 +13,17 @@ The solve overwrites the residual with the defect.
 """
 function IRTriangle!(AF::Union{MPLFact,MPHFact}, r, rs, verbose)
     AFS = AF.AF
-    on_the_fly=AF.onthefly
+    on_the_fly = AF.onthefly
     if on_the_fly
-    ldiv!(AFS, r)
+        ldiv!(AFS, r)
     else
-#        TFact = eltype(AFS)
-#        TW = eltype(r)
-#        rs .= TFact.(r)
+        #        TFact = eltype(AFS)
+        #        TW = eltype(r)
+        #        rs .= TFact.(r)
         rs .= r
         ldiv!(AFS, rs)
         r .= rs
-#r .= TW.(AFS\rs)
+        #r .= TW.(AFS\rs)
     end
     return r
 end
