### 2024-09-24 1fa2ac7
code cleanup part 1
