### 2024-09-24 980c682
code cleanup phase 2
