### 2024-09-24 0ef4538
Update paper; remove example
