### 2024-09-24 #6698 @ctkelley
Hi @rafaelbailo <https://github.com/rafaelbailo> . Thanks again for the
thoughtful review. Here is a point-by-point reply to your report. I have
mostly followed your suggestions. The exception is the "no PR please"
statement in the readme. I have explained that in my response

-- Tim

The Paper

   1. Code examples in JOSS papers: I eliminated that section as you
   suggested.

The Repository

   1.

   link to documentation in About section: Done
   2.

   LinearAlgebra.jl: I used your idea about Reexport.jl
   3.

   Dependencies section: I reworded the part about "install"
   4.

   Community guidelines:

   A) Issue templates: Done

   B) I would like to avoid PRs until I get QR and
   nonlinear solvers finished sometime in the next 6-12 months. These
   two things are tricky and I do not want to have to deal with PRs at
   the same time. After that's done, PRs will be welcome.

The Code:

   1.

   LinearAlgebra.jl alternative: Done
   2.

   CodeCov has done that to me before. I think it's harmless.
   3.

   Stray comments: Cleaned up
   4.

   Formatting: Done
   5.

   Docs: fixed in item 1 of Repository

On Mon, Sep 23, 2024 at 12:31 PM Rafael Bailo ***@***.***>
wrote:

> Hi @ctkelley <https://github.com/ctkelley> (CC: @vissarion
> <https://github.com/vissarion>), I have just submitted my review comments
> as an issue
> <https://github.com/ctkelley/MultiPrecisionArrays.jl/issues/51> in the
> package's repository. Please let me know when you get the chance to address
> them, or if you'd like to discuss any of them further.
>
