### 2024-09-25 #6698 @rafaelbailo
Hi @vissarion, @ctkelley has addressed all of [my suggestions](https://github.com/ctkelley/MultiPrecisionArrays.jl/issues/51) and I have ticked every item on my reviewer checklist. As far as I'm concerned, the submission is ready for acceptance. 👍🏻
