### 2024-09-25 ctkelley/MultiPrecisionArrays.jl#60 @vissarion
JOSS references
This is part of https://github.com/openjournals/joss-reviews/issues/6698

1. Some DOIs are missing. e.g.

- Improving the Accuracy of Computed Eigenvalues and Eigenvectors  - https://doi.org/10.1137/0720002
- Bi-CGSTAB: A Fast and Smoothly Converging Variant of Bi-CG for the Solution of Nonsymmetric Linear Systems - https://doi.org/10.1137/0913035

2. I think the following ref is not needed Kelley, C. T. (2024b). MultiPrecisionArrays.jl. https://github.com/ctkelley/MultiPrecisionArrays.jl. https://doi.org/10.5281/zenodo.7521427

3. In ref Kelley, C. T. (2022c) the journal title/pages/issue etc are missing.
