### 2024-09-25 #6698 @vissarion
Thanks @rafaelbailo for your quick review!

@ctkelley I have opened an issue regarding references https://github.com/ctkelley/MultiPrecisionArrays.jl/issues/60 and also a PR that fixes some typos https://github.com/ctkelley/MultiPrecisionArrays.jl/pull/59 please let me know if this is OK.
