### 2024-09-25 fbb3bce
update bib
