+++ Joss_Paper/paper.bib
@@ -38,7 +38,8 @@ title="Improving the accuracy of computed eigenvalues and eigenvectors",
 journal="SIAM Journal on Numerical Analysis",
 volume=20,
 pages="23--45",
-year=1983
+year=1983,
+doi={10.1137/0720002}
 }
 
 @techreport{Wilkinson:48,
@@ -155,7 +156,8 @@ title="{Bi-CGSTAB}: A fast and smoothly converging variant to
 journal = {SIAM J. Sci. Stat. Comp.},
 year=1992, 
 pages="631--644",
-volume=13
+volume=13,
+doi={10.1137/0913035}
 }
 
 @ARTICLE{saad:1986,
