### 2024-09-25 ctkelley/MultiPrecisionArrays.jl#60 @ctkelley
Hi,

1) I fixed the DOIs
2) That ref is for the package so the link will be in the paper. Is that a
mistake?
3) kelley:2022c is not a journal article. It is a Julia package.

— Tim

On Wed, Sep 25, 2024 at 5:21 AM Vissarion Fisikopoulos <
***@***.***> wrote:

> This is part of openjournals/joss-reviews#6698
> <https://github.com/openjournals/joss-reviews/issues/6698>
>
>    1. Some DOIs are missing. e.g.
>
>
>    - Improving the Accuracy of Computed Eigenvalues and Eigenvectors -
>    https://doi.org/10.1137/0720002
>    - Bi-CGSTAB: A Fast and Smoothly Converging Variant of Bi-CG for the
>    Solution of Nonsymmetric Linear Systems -
>    https://doi.org/10.1137/0913035
>
>
>    2.
>
>    I think the following ref is not needed Kelley, C. T. (2024b).
>    MultiPrecisionArrays.jl.
>    https://github.com/ctkelley/MultiPrecisionArrays.jl.
>    https://doi.org/10.5281/zenodo.7521427
>    3.
>
>    In ref Kelley, C. T. (2022c) the journal title/pages/issue etc are
>    missing.
>
