### 2024-09-25 ctkelley/MultiPrecisionArrays.jl#60 @ctkelley
kelley:2022c in the bib file is a Julia package
kelley:2022c in the paper is as book

I do not understand why this happened.
