### 2024-09-25 ctkelley/MultiPrecisionArrays.jl#51 @ctkelley
Thanks again.

— Tim

On Wed, Sep 25, 2024 at 3:27 AM Rafael Bailo ***@***.***>
wrote:

> Hi @ctkelley <https://github.com/ctkelley>, thank you for the prompt
> response and all the work! Fair enough about the PRs, I completely
> understand that you don't want to juggle PRs while working on core
> components of the package. I'm closing this as completed and recommending
> acceptance on the review issue. 👍🏻
>
