### 2024-09-25 bcc3be4
reword "iteration loop"
