### 2024-09-25 ctkelley/MultiPrecisionArrays.jl#60 @ctkelley
I fixed the "iteration loop" wording.
