### 2024-09-25 ctkelley/MultiPrecisionArrays.jl#60 @ctkelley
OK. When github made the paper it swapped 2022b and 2022c from their order
in the .bib file. No harm done.
