### 2024-09-26 ctkelley/MultiPrecisionArrays.jl#60 @vissarion
2. Yes, it is a reference to the package under submission which eventually will be published and get a DOI (and also a zenodo link) I think it is confusing to have two links and two DOIs for the same package [MultiPrecisionArrays.jl](https://github.com/ctkelley/MultiPrecisionArrays.jl).
3. kelley:2022c is a book in the pdf, the following information are from SIAM, could be added but I think it is OK as is since it is a book
```
Published:2022
ISBN:978-1-61197-726-4
eISBN:978-1-61197-727-1
https://doi.org/10.1137/1.9781611977271
Book Series Name:Fundamentals of Algorithms
```
