### 2024-09-26 c2e050c
bib fixes
