+++ Joss_Paper/paper.bib
@@ -79,12 +79,13 @@ journal="SIAM Review"
 author="C. T. Kelley",
 title="{Solving Nonlinear Equations with Iterative Methods:
 Solvers and Examples in Julia}",
-year=2022,
-publisher="SIAM",
+publisher="SIAM, Philadelphia",
 address="Philadelphia",
-isbn="978-1-611977-26-4",
-series="Fundamentals of Algorithms",
-doi={10.1137/1.9781611977271},
+Published:2022
+ISBN:978-1-61197-726-4
+eISBN:978-1-61197-727-1
+https://doi.org/10.1137/1.9781611977271
+Book Series Name:Fundamentals of Algorithms
 number=20
 }
 
