+++ Joss_Paper/paper.bib
@@ -81,11 +81,11 @@ title="{Solving Nonlinear Equations with Iterative Methods:
 Solvers and Examples in Julia}",
 publisher="SIAM, Philadelphia",
 address="Philadelphia",
-Published:2022
-ISBN:978-1-61197-726-4
-eISBN:978-1-61197-727-1
-https://doi.org/10.1137/1.9781611977271
-Book Series Name:Fundamentals of Algorithms
+Published:2022,
+ISBN:978-1-61197-726-4,
+eISBN:978-1-61197-727-1,
+https://doi.org/10.1137/1.9781611977271,
+Book Series Name:Fundamentals of Algorithms,
 number=20
 }
 
