### 2024-09-26 a309c5a
fix bib bugs
