+++ Joss_Paper/paper.bib
@@ -85,7 +85,7 @@ year=2022,
 ISBN ="978-1-61197-726-4",
 eISBN = "978-1-61197-727-1",
 doi="10.1137/1.9781611977271",
-Book Series Name="Fundamentals of Algorithms",
+series="Fundamentals of Algorithms",
 number=20
 }
 
