### 2024-09-26 e51b9b4
bib fixes
