### 2024-09-26 ctkelley/MultiPrecisionArrays.jl#60 @ctkelley
I've eliminated the reference to the package.

I tried your changes to the citation to the book and got errors from GitHub actions. I'll stick with what I had if that is ok.
