### 2024-09-27 ctkelley/MultiPrecisionArrays.jl#60 @vissarion
Thanks for the efforts, references look good now!
