### 2024-09-27 #6698 @ctkelley
Hi, I released and tagged v0.1.3, but am unable to update the Zenodo DOI.
When I get to Zenodo it asks me to make a new release. That is what I just
did.

What do I do now?

I had a DOI for the repo but I am not clear on how to update that one.

On Fri, Sep 27, 2024 at 5:11 AM Vissarion Fisikopoulos <
***@***.***> wrote:

> I am also OK after parsing the text and all my comments are addressed.
>
> @ctkelley <https://github.com/ctkelley> when a submission is ready to be
> accepted, we ask that the authors issue a new tagged release of the
> software (if changed), and archive it (see this guide
> <https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content>).
> Please do this and post the version number and archive DOI here.
> Please make sure that the author names, affiliations as well as the title
> of the archive (e.g. zenodo) is exactly the same with the submission.
>
