### 2024-09-27 #6698 @ctkelley
I think I know what's going on. I got a Zenodo DOI for the package a couple
years ago and then renamed it. Zendo does not seem to be able to handle
that. When I registered v0.1.3 with JuliaHub TagBot automatically made it
release 0.1.3 via GitHub actions.

So, Zenodo will let me get a new DOI, but I will have to make a new tag.
So, is it ok if I tag v0.1.3.1 but leave the package version a 0.1.3 to be
consistent with the Julia registry?
