+++ v0.1.3-joss/siteinfo.js
@@ -1,0 +1,1 @@
+var DOCUMENTER_CURRENT_VERSION = "v0.1.3-joss";
