### 2024-09-27 #6698 @ctkelley
I got it to work with a tag for  v0.1.3-joss.

Version = 0.1.3
DOI = 10.5281/zenodo.13851500

Let me know if this is ok.
