+++ Project.toml
@@ -1,26 +1,0 @@
-name = "MultiPrecisionArrays"
-uuid = "e0585029-f0fd-471c-997a-b805c9f71f28"
-authors = ["C. T. (Tim) Kelley <tim_kelley@ncsu.edu>"]
-version = "0.1.3"
-
-[deps]
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-OhMyThreads = "67456a42-1dca-4109-a031-0a68de7e3ad5"
-Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-SIAMFANLEquations = "084e46ad-d928-497d-ad5e-07fa361a48c4"
-SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-
-[compat]
-LinearAlgebra = "1"
-OhMyThreads = "0.5, 0.6"
-Reexport = "1.2.2"
-SIAMFANLEquations = "1"
-SparseArrays = "1"
-Test = "1"
-julia = "1"
-
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Test"]
