+++ Project.toml
@@ -1,7 +1,7 @@
 name = "MultiPrecisionArrays"
 uuid = "e0585029-f0fd-471c-997a-b805c9f71f28"
 authors = ["C. T. (Tim) Kelley <tim_kelley@ncsu.edu>"]
-version = "0.1.3"
+version = "0.1.4"
 
 [deps]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
