+++ dev/MPArray.bib
@@ -184,7 +184,7 @@ title="{MultiPrecisionArrays.jl}",
 author="C. T. Kelley",
 year=2024,
 note="Julia Package",
-doi="10.5281/zenodo.7521427",
+doi="10.5281/zenodo.13851500",
 url="https://github.com/ctkelley/MultiPrecisionArrays.jl"
 }
 
@@ -232,5 +232,12 @@ archivePrefix={arXiv},
 primaryClass={math.NA},
 eprint={2407.00827},
 note="submitted for publication"
+}
+
+@misc{ctk:joss2,
+author="C. T. Kelley",
+title="{MultiPrecisionArrays.jl}: A {J}ulia package for iterative refinement",
+year=2024,
+note="Submitted to Journal of Open Source Software"
 }
 
