### 2024-09-30 #6698 @vissarion
Thanks @ctkelley!

We need one more change in the zenodo title. "Please make sure that the author names, affiliations as well as the title of the archive (e.g. zenodo) is exactly the same with the submission."
