### 2024-09-30 #6698 @ctkelley
Proofs look good to me. Missing DOIs are (1) paper from 1948 and (2) paper under review.
