### 2024-09-30 de35413
minor changes in paper.md
