+++ Joss_Paper/paper.bib
@@ -33,7 +33,7 @@ doi = {10.1137/17M1140819}
 }
 
 @article{dongarra:1983,
-author="J.J.Dongarra and C.B.Moler and J.H.Wilkinson",
+author="J.J. Dongarra and C.B. Moler and J.H. Wilkinson",
 title="Improving the accuracy of computed eigenvalues and eigenvectors",
 journal="SIAM Journal on Numerical Analysis",
 volume=20,
