### 2024-09-30 #6698 @danielskatz
@ctkelley - As an AEiC responsible for the next steps, I've suggested a bunch of small changes in https://github.com/ctkelley/MultiPrecisionArrays.jl/pull/69 - please merge this, or let me know what you disagree with, then we can continue the acceptance and publication process.
