### 2024-09-30 #6698 @ctkelley
Merged! Thanks for your help with this paper.

— Tim

On Mon, Sep 30, 2024 at 9:58 AM Daniel S. Katz ***@***.***>
wrote:

> @ctkelley <https://github.com/ctkelley> - As an AEiC responsible for the
> next steps, I've suggested a bunch of small changes in
> ctkelley/MultiPrecisionArrays.jl#69
> <https://github.com/ctkelley/MultiPrecisionArrays.jl/pull/69> - please
> merge this, or let me know what you disagree with, then we can continue the
> acceptance and publication process.
>
