### 2024-09-30 #6698 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
