### 2024-09-30 #6698 @ctkelley
pdf looks good

On Mon, Sep 30, 2024 at 11:01 AM The Open Journals editorial robot <
***@***.***> wrote:

> 👋 @openjournals/csism-eics
> <https://github.com/orgs/openjournals/teams/csism-eics>, this paper is
> ready to be accepted and published.
>
> Check final proof 👉📄 Download article
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.06698/joss.06698/10.21105.joss.06698.pdf>
>
> If the paper PDF and the deposit XML files look good in
> openjournals/joss-papers#5935
> <https://github.com/openjournals/joss-papers/pull/5935>, then you can now
> move forward with accepting the submission by compiling again with the
> command @editorialbot accept
>
