### 2024-09-30 #6698 @danielskatz
👋 Congratulations to @ctkelley (Carl Kelley) on your publication!!

JOSS depends on volunteers and couldn't be successful without you
